Abstracted into quoteCreator utility and integrated into init.
