+++ src/controllers/b2b/bpp/init.ts
@@ -1,13 +1,29 @@
 import { Request, Response } from "express";
 import { onInitDomestic } from "../../../lib/examples";
-import { ACTIONS, responseBuilder } from "../../../lib/utils";
+import { ACTIONS, quoteCreator, responseBuilder } from "../../../lib/utils";
 
 export const initController = (req: Request, res: Response) => {
+	const { context, message } = req.body;
+	const { items, fulfillments, tags, billing, ...remainingMessage } =
+		message.order;
+	const {type, collected_by, ...staticPaymentInfo} = onInitDomestic.message.order.payments[0];
+	const responseMessage = {
+		order: {
+			items,
+			fulfillments,
+			tags,
+			billing,
+			provider: {id: remainingMessage.provider.id},
+			provider_location: remainingMessage.provider.locations[0],
+			payments: remainingMessage.payments.map((each: any) => ({...each, ...staticPaymentInfo})),
+			quote: quoteCreator(items)
+		},
+	};
 	return responseBuilder(
 		res,
-		req.body.context,
-		onInitDomestic.message,
-		req.body.context.bap_uri,
-    `on_${ACTIONS.init}`
+		context,
+		responseMessage,
+		context.bap_uri,
+		`on_${ACTIONS.init}`
 	);
 };
