+++ src/lib/utils/responseBuilder.ts
@@ -3,6 +3,7 @@ import { Response } from "express";
 import { v4 as uuidv4 } from "uuid";
 import { MOCKSERVER_ID, MOCKSERVER_URL } from "./constants";
 import { createResponseAuthHeader } from "./responseAuth";
+import { onSelectDomestic } from "../examples";
 
 export const responseBuilder = async (
 	res: Response,
@@ -81,4 +82,59 @@ export const responseBuilder = async (
 			async,
 		});
 	}
+};
+
+export const quoteCreator = (
+	items: typeof onSelectDomestic.message.order.items
+) => {
+	var breakup: any[] = [];
+	const onFulfillment = onSelectDomestic.message.order.quote.breakup.filter(
+		(each) => each["@ondc/org/item_id"] === "F1"
+	);
+	const onItem = onSelectDomestic.message.order.quote.breakup.filter(
+		(each) =>
+			each["@ondc/org/item_id"] === "I1" &&
+			each["@ondc/org/title_type"] !== "item"
+	);
+	items.forEach((item: any) => {
+		breakup = [
+			...breakup,
+			...onItem,
+			{
+				"@ondc/org/item_id": item.id,
+				"@ondc/org/item_quantity": {
+					count: item.quantity.selected.count,
+				},
+				title: "Product Name Here",
+				"@ondc/org/title_type": "item",
+				price: {
+					currency: "INR",
+					value: (item.quantity.selected.count * 250).toString(),
+				},
+				item: {
+					price: {
+						currency: "INR",
+						value: "250",
+					},
+				},
+			},
+		];
+		item.fulfillment_ids.forEach((eachId: string) => {
+			breakup = [
+				...breakup,
+				...onFulfillment.map((each) => ({
+					...each,
+					"@ondc/org/item_id": eachId,
+				})),
+			];
+		});
+	});
+	return {
+		breakup,
+		price: {
+			currency: "INR",
+			value: (53_600 * items.length).toString(),
+		},
+		ttl: "P1D",
+	};
 };
